pyproject.toml
[project]
name = "hmeqo"
version = "0.1.0"
description = ""
authors = [{ name = "hmeqo", email = "hmeqocoliniliad@gmail.com" }]
dependencies = [
    "opencv-python>=4.8.1.78",
    "numpy>=1.25.2",
    "django>=4.2.6",
    "pillow>=10.1.0",
    "django-cors-headers>=4.3.0",
    "djangorestframework>=3.14.0",
    "drf-yasg>=1.21.7",
    "django-debug-toolbar>=4.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
